Delegate Subscriptions
Created events to pass in current shape selection to player projectile, so that upon activation, the projectile has a different sprite and enum identity.

Prevents active objects from changing shape once activated with a bool check that means the logic can only run on inactive projectiles. This is due to all projectiles being subbed to the event, meaning it is called on all whether active or not.