Added all enemy animations
Animations have been sprite sheeted and added to an animation controller.

Need to make enemy have anims set based on triggers that will be triggered from their "Shape Info " file upon creation from an object pool.

For example, if the shape type is "Triangle", this will be passed into a method that sets the trigger of "isTriangle" upon object creation.