Quick Null Exception fix
Fixed Null Exception in Player_Bullet_Object_Pool.

Issue was an event from the on screen UI Manager (For displaying the player's current shape) not subscribing to an event in the object pool in the correct order, leading to an event being fired whilst having no listeners.

Prevented this by not calling the event so early in the start method, and instead just hard coding default values for the shape that the player will start with.